requirements.txt
PyQt5==5.15.9
Pillow>=10.0.0
numpy>=1.26.0
